hillclimber step 1
