+++ objects/area.py
@@ -1,3 +1,5 @@
+import random
+
 class Area(object):
     width = 320
     height = 360
@@ -13,19 +15,22 @@ class Area(object):
         house.x = x
         house.y = y
         kind = type(house).__name__
-        if house.check_validity():
-            # place the house on every coordinate that is covered by the house
-            for i in range(x, x + house.width):
-                for j in range(y, y + house.height):
-                    self.grid[i][j] = house
-            # add the house to the appropriate list
-            if kind == "Mansion":
-                self.mansionList.append(house)
-            elif kind == "Bungalow":
-                self.bungalowList.append(house)
-            elif kind == "FamilyHome":
-                self.familyHomeList.append(house)
-        else:
+
+        # place new house
+        try:
+            if house.check_validity():
+                # place the house on every coordinate that is covered by the house
+                for i in range(x, x + house.width):
+                    for j in range(y, y + house.height):
+                        self.grid[i][j] = house
+                # add the house to the appropriate list
+                if kind == "Mansion":
+                    self.mansionList.append(house)
+                elif kind == "Bungalow":
+                    self.bungalowList.append(house)
+                elif kind == "FamilyHome":
+                    self.familyHomeList.append(house)
+        except:
             raise RuntimeError("Cannot validly place "
                                "house at these coordinates.")
 
@@ -72,3 +77,93 @@ class Area(object):
             totalPrice += self.familyHomeList[i].get_price()
             i += 1
         return totalPrice
+
+             
+    def sliding_house(self, currentHouse, backupX, backupY):
+        # remove house from map
+        self.remove_house(currentHouse)
+
+        directionShift = None
+
+        # determine distance to move and update house coordinates 
+        currentHouse = self.determineShift(currentHouse, directionShift)
+
+        if not self.place_house(currentHouse, currentHouse.x, currentHouse.y): 
+            # place origanal house
+            currentHouse.x = backupX
+            currentHouse.y = backupY
+            self.place_house(currentHouse, currentHouse.x, currentHouse.y)
+
+    def determineShift(self, currentHouse, directionShift):
+        
+        # choose to move horizontal or vertical
+        if directionShift == None:
+            directionShift = random.randint(0, 1)
+            print("Direction: {}".format(directionShift))
+
+        # pick random distance to shift the house with
+        amountShift = random.randint(-10, 10)
+        print("amountShift: {}".format(amountShift))
+
+        # move house in chosen direction,
+        # but only if it still falls within the map TODO break maken
+        recursiveCount = 0
+        if directionShift == 0:
+            tempCurrentHouseX = currentHouse.x + amountShift
+            tempBoundry = (self.width
+                           - currentHouse.width
+                           - currentHouse.minimumSpace)
+            if (tempCurrentHouseX > currentHouse.minimumSpace and
+                    tempCurrentHouseX < tempBoundry):
+                currentHouse.x += amountShift
+                return currentHouse
+            else:
+                recursiveCount += 1
+                print("AmountShift ({}) not possible "
+                      "(house would be outside map)".format(amountShift))
+                if recursiveCount < 50:
+                    secondRecursiveCount += 1
+                    # change directoin horizontal vertical
+                    if secondRecursiveCount < 2:
+                        if directionShift == 1:
+                            directionShift = 0
+                            self.determineShift(currentHouse, directionShift)
+                        else:
+                            directionShift = 1
+                            self.determineShift(currentHouse, directionShift)   
+                    else:
+                        print("return currenthouse")
+                        return currentHouse
+                    
+        else:
+            tempCurrentHouseY = currentHouse.y + amountShift
+            tempBoundry = (self.height
+                           - currentHouse.height
+                           - currentHouse.minimumSpace)
+            if (tempCurrentHouseY > currentHouse.minimumSpace and
+                    tempCurrentHouseY < tempBoundry):
+                currentHouse.y += amountShift
+                return currentHouse
+            else:
+                print("AmountShift ({}) not possible "
+                      "(house would be outside map)".format(amountShift))
+                self.determineShift(currentHouse, directionShift)
+
+        # recursive error catching
+        # returning currenthouse from last valid determineShift attempt
+        return currentHouse
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
